add xml support
