+++ Sources/OracleNIO/Constants.swift
@@ -68,6 +68,7 @@ enum Constants {
     static let TNS_OBJ_IS_VERSION_81 = 0x80
     static let TNS_OBJ_IS_DEGENERATE = 0x10
     static let TNS_OBJ_IS_COLLECTION = 0x08
+    @usableFromInline
     static let TNS_OBJ_NO_PREFIX_SEG: UInt8 = 0x04
     static let TNS_OBJ_IMAGE_VERSION = 1
 
@@ -103,8 +104,11 @@ enum Constants {
     static let TNS_OBJ_TDS_TYPE_BINARY_DOUBLE = 45
 
     // MARK: XML type constants
+    @usableFromInline
     static let TNS_XML_TYPE_LOB: UInt32 = 0x0001
+    @usableFromInline
     static let TNS_XML_TYPE_STRING: UInt32 = 0x0004
+    @usableFromInline
     static let TNS_XML_TYPE_FLAG_SKIP_NEXT_4: UInt32 = 0x100000
 
     // MARK: Execute options
